+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Introduction of a new module.
 
 - BREAKING FEATURE: `child.get_screenshot()` now by default calls `:redraw` prior to computing screenshot. Can be disabled by new `opts.redraw` argument.
 - FEATURE: new method `child.lua_func()` can execute simple functions inside child process and return the result (stasjok, #437).
+- FEATURE: `expect.reference_screenshot()` now has `ignore_lines` option allowing to ignore specified lines during screenshot compare.
 
 
 # Version 0.10.0